refactor: extract argument-modification.js out
